+++ packages/gelly/modules/gelly-fluid-renderer/src/v2/renderers/splatting/splatting-renderer.h
@@ -5,6 +5,7 @@
 #include "GellyInterfaceRef.h"
 #include "device.h"
 #include "fluidsim/ISimData.h"
+#include "helpers/rendering/gpu-duration.h"
 #include "pipeline/pipeline.h"
 #include "pipelines/pipeline-info.h"
 #include "renderdoc_app.h"
@@ -38,6 +39,18 @@ class SplattingRenderer {
 public:
 	float ALBEDO_OUTPUT_SCALE = 0.25f;
 
+	/**
+	 * Timings of all the passes in milliseconds.
+	 */
+	struct Timings {
+		float ellipsoidSplatting = 0.0f;
+		float albedoDownsampling = 0.0f;
+		float surfaceFiltering = 0.0f;
+		float rawNormalEstimation = 0.0f;
+
+		bool isDisjoint = false;
+	};
+
 	struct Settings {
 		unsigned int filterIterations = 5;
 		/**
@@ -49,6 +62,14 @@ public:
 		 */
 		bool enableGPUSynchronization = true;
 		bool enableSurfaceFiltering = true;
+		/**
+		 * Should never be true for general use, but this is useful for
+		 * debugging on a user's machine. This will enable the GPU timing
+		 * queries, which will allow for coarse-grained GPU timing information
+		 * to be captured. It is way more in-depth than a simple CPU timing, way
+		 * less than a GPU profiler.
+		 */
+		bool enableGPUTiming = false;
 	};
 
 	struct SplattingRendererCreateInfo {
@@ -72,6 +93,8 @@ public:
 	auto SetFrameResolution(float width, float height) -> void;
 	auto GetSettings() const -> Settings;
 	auto UpdateSettings(const Settings &settings) -> void;
+	auto FetchTimings() -> Timings;
+
 	[[nodiscard]] auto GetAbsorptionModifier() const
 		-> std::shared_ptr<AbsorptionModifier>;
 
@@ -95,6 +118,15 @@ private:
 #ifdef GELLY_ENABLE_RENDERDOC_CAPTURES
 	RENDERDOC_API_1_1_2 *renderDoc = nullptr;
 #endif
+
+	struct {
+		util::GPUDuration ellipsoidSplatting;
+		util::GPUDuration albedoDownsampling;
+		util::GPUDuration surfaceFiltering;
+		util::GPUDuration rawNormalEstimation;
+	} durations;
+
+	Timings latestTimings;
 
 	auto CreatePipelines() -> void;
 	auto CreatePipelineInfo() const -> PipelineInfo;